[DOC-10] Fix links in PR template
Problem: apaprently GitHub uses different rules for markdown links when
it displays them in a PR. For instance, `/README.md` is treated as
`github.com/README`. Probably it's a GitHub's bug, but that's how it works now.
Solution: prefix links with `../tree/master/` which somehow works. That looks
like a workaround for a bug, but I am not sure, maybe it's intentional.