Update "Fixed #1" comment in PR template
Problem: some services (e. g. gitlab.com) treat "Fixed #1" comment
as a statement that this PR/MR fixes issue #1 (even though it's a
comment). Because of that each PR which uses that template may be
considered as a fix for #1.
Solution: replace it with #100500 and hope that no repo will have
so many issues.