[PBP-86] Add licensing headers
Problem: according to our current best practices we want each text
file to contain a header which specifies its licensing information.
However, files in this repo do not have these headers.

Solution: add headers to all files that are supposed to be copied
to a new repo when it is created. Exception: files in `reuse` which
apparently do not have to have these headers.