[PBP-86] Update Makefile template
Problem: our Makefile template was copied by me from somewhere and
as far as I know it is not used anywhere.
We have come up with a different Makefile in Morley.
Another (probably minor) problem/inconvenience with this Makefile
template is that we do not own its copyright.
Solution: update Makefile using one from Morley.