Write about Resolves keyword in RP template
Problem: GitHub and GitLab have special keywords to close an issue when
a PR is merged, but our PR template doesn't encourage their usage.
Most commonly a PR closes an issue, so it makes sense to use a
corresponding keyword.
Solution: add a meta-comment that asks to add "Resolves" prefix for
GitHub/GitLab issue trackers.
Also ask to use short link in this case because it's more stable.