[INT-250] Add `Fixed` to the PR template
Problem: Sometimes, devs fill in the "Related issue(s)" section of the
PR template by simply writing down the issue number, e.g. `#123`, and
they forget to prefix it with the word "Fixed". As a result, when the PR
is merged, github won't automatically close the issue.

Solution: Add the "Fixed #" prefix directly to the template.