[INT-250] Add release checklist to PR template
Problem: Each repo usually has a slightly different release process, and
some of those steps tend to be forgotten when working on multiple repos.
Recently, we've been adding "release checklists" to each repo to
document this process and make sure steps are not forgotten.

Solution: Add a release checklist to the PR template in the
metatemplates project.